Close connection, even if "username" was in kwargs This fixes a resource usage warning in the async unittests.

+++ tests/test_asyncio/conftest.py
@@ -99,9 +99,7 @@ async def create_redis(request):
 
         async def teardown():
             if not cluster_mode:
-                if "username" in kwargs:
-                    return
-                if flushdb:
+                if flushdb and "username" not in kwargs:
                     try:
                         await client.flushdb()
                     except redis.ConnectionError:
